ISPN-3854 Tests from org.infinispan.eviction package fail randomly with JDK7 and constantly with IBM JDK6
* always lock when load from persistence and put in memory
