+++ core/src/main/java/org/infinispan/interceptors/ActivationInterceptor.java
@@ -1,19 +1,15 @@
 package org.infinispan.interceptors;
 
 import org.infinispan.commands.FlagAffectedCommand;
-import org.infinispan.commands.write.ApplyDeltaCommand;
 import org.infinispan.commands.write.PutMapCommand;
 import org.infinispan.configuration.cache.Configuration;
 import org.infinispan.configuration.cache.EvictionConfiguration;
-import org.infinispan.container.DataContainer;
 import org.infinispan.container.entries.CacheEntry;
-import org.infinispan.container.entries.InternalCacheEntry;
 import org.infinispan.context.InvocationContext;
 import org.infinispan.eviction.ActivationManager;
 import org.infinispan.eviction.EvictionStrategy;
 import org.infinispan.factories.annotations.Inject;
 import org.infinispan.factories.annotations.Start;
-import org.infinispan.interceptors.locking.ClusteringDependentLogic;
 import org.infinispan.util.logging.Log;
 import org.infinispan.util.logging.LogFactory;
 
@@ -26,8 +22,6 @@ public class ActivationInterceptor extends CacheLoaderInterceptor {
    private Configuration cfg;
    private boolean isManualEviction;
    private ActivationManager activationManager;
-   private ClusteringDependentLogic cdl;
-   private DataContainer dataContainer;
 
    @Override
    protected Log getLog() {
@@ -35,12 +29,9 @@ public class ActivationInterceptor extends CacheLoaderInterceptor {
    }
 
    @Inject
-   public void inject(Configuration cfg, ActivationManager activationManager, ClusteringDependentLogic cdl,
-                      DataContainer dataContainer) {
+   public void inject(Configuration cfg, ActivationManager activationManager) {
       this.cfg = cfg;
       this.activationManager = activationManager;
-      this.cdl = cdl;
-      this.dataContainer = dataContainer;
    }
 
    @Start(priority = 15)
@@ -65,41 +56,9 @@ public class ActivationInterceptor extends CacheLoaderInterceptor {
    }
 
    @Override
-   protected Boolean loadIfNeeded(InvocationContext ctx, Object key, boolean isRetrieval, FlagAffectedCommand cmd) throws Throwable {
-      CacheEntry entry = ctx.lookupEntry(key);
-      if (!shouldAttemptLookup(entry)) {
-         //the entry is already in the context. Avoid look in the cache loader.
-         return null;
-      }
-      try {
-         while (!cdl.lock(key, false));
-         InternalCacheEntry ice = dataContainer.get(key);
-         if (ice != null) {
-            if (cmd instanceof ApplyDeltaCommand) {
-               ctx.putLookedUpEntry(key, ice);
-               entryFactory.wrapEntryForDelta(ctx, key, ((ApplyDeltaCommand) cmd).getDelta());
-            } else {
-               entryFactory.wrapEntryForPut(ctx, key, ice, false, cmd, false);
-            }
-            return null;
-         }
-         Boolean loaded = super.loadIfNeeded(ctx, key, isRetrieval, cmd);
-         if (loaded == Boolean.TRUE) {
-            if (enabled && isManualEviction) {
-               // check if value was loaded
-               CacheEntry e = ctx.lookupEntry(key);
-               // We have to commit this before we remove from the store so there isn't a gap where the entry isn't visible
-               // This should always be true given that loaded was true, but just sanity check
-               if (e != null && e.isLoaded()) {
-                  cdl.commitEntry(e, null, cmd, ctx);
-                  removeFromStoreIfNeeded(key);
-               }
-            }
-         }
-         return loaded;
-      } finally {
-         cdl.unlock(key);
-      }
+   protected void afterSuccessfullyLoaded(CacheEntry wrappedEntry) {
+      //the method already checks for enable and isManualEviction.
+      removeFromStoreIfNeeded(wrappedEntry.getKey());
    }
 
    @Override
