+++ core/src/test/java/org/infinispan/eviction/ManualEvictionWithConcurrentOperationsInBackupOwnerTest.java
@@ -1,32 +1,21 @@
 package org.infinispan.eviction;
 
-import org.infinispan.Cache;
-import org.infinispan.distribution.DistributionManager;
-import org.infinispan.remoting.transport.Address;
+import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.testng.annotations.Test;
 
 /**
- * Tests manual eviction with concurrent read and/or write operation. This test has passivation disabled and the
- * eviction happens in the backup owner
+ * Same as {@link ManualEvictionWithSizeBasedAndConcurrentOperationsInBackupOwnerTest} but with an unbounded data
+ * container
  *
  * @author Pedro Ruivo
  * @since 6.0
  */
 @Test(groups = "functional", testName = "eviction.ManualEvictionWithConcurrentOperationsInBackupOwnerTest", singleThreaded = true)
 public class ManualEvictionWithConcurrentOperationsInBackupOwnerTest
-      extends ManualEvictionWithConcurrentOperationsInPrimaryOwnerTest {
+      extends ManualEvictionWithSizeBasedAndConcurrentOperationsInBackupOwnerTest {
 
    @Override
-   protected Object createSameHashCodeKey(String name) {
-      final Cache otherCache = otherCacheManager.getCache();
-      final Address address = otherCache.getAdvancedCache().getRpcManager().getAddress();
-      DistributionManager distributionManager = otherCache.getAdvancedCache().getDistributionManager();
-      int hashCode = 0;
-      SameHashCodeKey key = new SameHashCodeKey(name, hashCode);
-      while (!distributionManager.getPrimaryLocation(key).equals(address)) {
-         hashCode++;
-         key = new SameHashCodeKey(name, hashCode);
-      }
-      return key;
+   protected void configureEviction(ConfigurationBuilder builder) {
+      builder.eviction().maxEntries(-1).strategy(EvictionStrategy.NONE);
    }
 }
